feat: Allowing not to remove user messages (#20)

+++ handlers/aliexpress_api_handler.py
@@ -12,7 +12,8 @@ from config import (
     ALIEXPRESS_TRACKING_ID,
     MSG_REPLY_PROVIDED_BY_USER,
     MSG_AFFILIATE_LINK_MODIFIED,
-    ALIEXPRESS_DISCOUNT_CODES
+    ALIEXPRESS_DISCOUNT_CODES,
+    DELETE_MESSAGES,
 )
 
 # Initialize logger
@@ -107,11 +108,17 @@ async def handle_aliexpress_api_links(message):
     if new_text != message.text:
         reply_to_message_id = message.reply_to_message.message_id if message.reply_to_message else None
         polite_message = f"{MSG_REPLY_PROVIDED_BY_USER} @{message.from_user.username}:\n\n{new_text}\n\n{MSG_AFFILIATE_LINK_MODIFIED}"
-        await message.delete()
-        logger.info(f"{message.message_id}: Original message deleted.")
 
-        await message.chat.send_message(text=polite_message, reply_to_message_id=reply_to_message_id)
-        logger.info(f"{message.message_id}: Sent modified message with affiliate links.")
+        if DELETE_MESSAGES:
+            # Deletes the original message and creates a new one
+            await message.delete()
+            await message.chat.send_message(text=polite_message, reply_to_message_id=reply_to_message_id)
+            logger.info(f"{message.message_id}: Original message deleted annd sent modified message with affiliate links.")
+        else:
+            # Replies to the original message without deleting it
+            await message.chat.send_message(text=polite_message, reply_to_message_id=message.message_id)
+            logger.info(f"{message.message_id}: Replied to the original message with affiliate links.")
+
         return True
 
     logger.info(f"{message.message_id}: No modifications made to the message.")
